consumer/requirements.txt
uuid
fastapi
asyncio
uvicorn[standard]
langchain
langchain_community
prometheus_fastapi_instrumentator
pydantic
openai
redis
opentelemetry-api
opentelemetry-sdk 
opentelemetry-exporter-jaeger 
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-requests
deprecated
protobuf>=3.20.0,<3.21.0
opentelemetry-exporter-otlp
PyPI
psycopg2-binary
aiokafka
asyncpg
pybreaker
fastapi-limiter[redis]